# openapi.yaml
openapi: 3.1.0
info:
  title: DreamWright API
  description: |
    AI-powered webtoon and short-form drama production API.

    ## Authentication
    All endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Workflow
    The typical workflow for creating a webtoon project is:
    1. Create a project (`POST /projects`)
    2. Expand a story prompt (`POST /projects/{id}/story`)
    3. Generate character assets (`POST /projects/{id}/characters/{id}/assets`)
    4. Generate location assets (`POST /projects/{id}/locations/{id}/assets`)
    5. Generate chapters sequentially (`POST /projects/{id}/chapters`)
    6. Generate panels for each chapter (`POST /projects/{id}/chapters/{id}/panels`)

    ## Asset Dependencies
    - Panel generation requires character portraits and location references to exist
    - Chapter N generation requires Chapter N-1 to exist (for story continuity)
    - Panel generation for Chapter N requires Chapter N-1 panels to exist (for visual continuity)

    ## Async Operations
    Long-running generation operations return `202 Accepted` with a job resource.
    Poll the job status endpoint or use webhooks for completion notifications.
  version: 1.0.0
  contact:
    name: DreamWright Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Projects]
      summary: Get project details
      operationId: getProject
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Projects]
      summary: Update project
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Projects]
      summary: Delete project
      operationId: deleteProject
      responses:
        '204':
          description: Project deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
...
  /projects/{projectId}/chapters/{chapterId}/panels:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/chapterId'

    get:
      tags: [Panels]
      summary: List panels for a chapter
      operationId: listPanels
      parameters:
        - name: scene_number
          in: query
          description: Filter by scene number
          schema:
            type: integer
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: List of panels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPanelList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Panels]
      summary: Generate panels for chapter
      description: |
        Generates panel images for all scenes in a chapter using AI.

        **Dependencies**:
        - All characters referenced in panels must have portrait assets
        - All locations referenced in scenes must have reference assets
        - Chapter N-1 must exist (for story context)
        - Chapter N-1 panels should exist (for visual continuity across chapters)

        Panels are generated sequentially within each scene to maintain visual continuity.

        This is an async operation.
      operationId: createPanels
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePanelsRequest'
      responses:
        '202':
          description: Panel generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Dependencies not met
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyError'
        '429':
          $ref: '#/components/responses/RateLimited'
        default:
          $ref: '#/components/responses/InternalError'
...
  /jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/jobId'

    get:
      tags: [Jobs]
      summary: Get job status
      description: Poll for the status of an async job
      operationId: getJob
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Jobs]
      summary: Cancel job
      description: Attempt to cancel a running job
      operationId: cancelJob
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Job cannot be cancelled (already completed or failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
...
    chapterId:
      name: chapterId
      in: path
      required: true
      description: Chapter ID (e.g., ch1, ch2)
      schema:
        type: string
...
    jobId:
      name: jobId
      in: path
      required: true
      description: Job ID for async operation
      schema:
        type: string
        format: uuid
    limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "unauthorized"
            message: "Authentication required"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "forbidden"
            message: "You do not have access to this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "rate_limited"
            message: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ProjectFormat:
      type: string
      enum: [webtoon, short_drama]
    ProjectStatus:
      type: string
      enum: [draft, in_progress, completed]
    Genre:
      type: string
      enum: [romance, action, fantasy, thriller, slice_of_life, horror, comedy, drama, mystery, scifi]
    Tone:
      type: string
      enum: [comedic, dramatic, dark, lighthearted, romantic, suspenseful]
    CharacterRole:
      type: string
      enum: [protagonist, antagonist, supporting, minor]
    LocationType:
      type: string
      enum: [interior, exterior]
...
    ShotType:
      type: string
      enum: [wide, medium, close_up, extreme_close_up]
    CameraAngle:
      type: string
      enum: [eye_level, high, low, dutch]
    DialogueType:
      type: string
      enum: [speech, thought, narration]
    ChapterStatus:
      type: string
      enum: [outlined, generating, completed]
    JobStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    ErrorCode:
      type: string
      enum:
        - bad_request
        - validation_error
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - dependency_not_met
        - asset_exists
        - rate_limited
        - internal_error
    DependencyType:
      type: string
      enum:
        - chapter
        - character_asset
        - location_asset
        - panel
        - story

    # Error schemas
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    DependencyError:
      type: object
      required: [code, message, missing_dependencies]
      properties:
        code:
          type: string
          enum: [dependency_not_met]
        message:
          type: string
        missing_dependencies:
          type: array
          items:
            $ref: '#/components/schemas/MissingDependency'
    MissingDependency:
      type: object
      required: [type, id, name, resolution]
      properties:
        type:
          $ref: '#/components/schemas/DependencyType'
        id:
          type: string
          description: Resource ID
        name:
          type: string
          description: Human-readable name
        resolution:
          type: string
          description: How to resolve this dependency
          example: "Generate character portrait: POST /projects/{id}/characters/{id}/assets"
...
    PaginationMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # Paginated lists
...
    PaginatedPanelList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Panel'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    Job:
      type: object
      required: [id, status, type, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        type:
          type: string
          description: Type of job (story_expansion, character_asset, panel_generation, etc.)
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Progress from 0.0 to 1.0
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        result:
          type: object
          nullable: true
          description: Result data when job completes
        error:
          $ref: '#/components/schemas/Error'
          nullable: true
...
    BatchJobResponse:
      type: object
      required: [job_id, status, poll_url, total_items]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        poll_url:
          type: string
          format: uri
        total_items:
          type: integer
          description: Number of items to process

    # Project schemas
...
    Project:
      type: object
      required: [id, name, format, status, created_at, characters, locations, chapters]
      properties:
        id:
          type: string
        name:
          type: string
        format:
          $ref: '#/components/schemas/ProjectFormat'
        status:
          $ref: '#/components/schemas/ProjectStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        story:
          $ref: '#/components/schemas/Story'
          nullable: true
        characters:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/ChapterSummary'
        original_prompt:
          type: string
...
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProjectStatus'
...
    StoryBeat:
      type: object
      required: [beat, description]
      properties:
        beat:
          type: string
        description:
          type: string
    Story:
      type: object
      required: [id, title, genre, tone, episode_count]
      properties:
        id:
          type: string
        title:
          type: string
        logline:
          type: string
        genre:
          $ref: '#/components/schemas/Genre'
        tone:
          $ref: '#/components/schemas/Tone'
        themes:
          type: array
          items:
            type: string
        target_audience:
          type: string
        episode_count:
          type: integer
        synopsis:
          type: string
        story_beats:
          type: array
          items:
            $ref: '#/components/schemas/StoryBeat'
...
    CharacterDescription:
      type: object
      properties:
        physical:
          type: string
        personality:
          type: string
        background:
          type: string
        motivation:
          type: string
    CharacterAssets:
      type: object
      properties:
        portrait:
          type: string
          nullable: true
          description: Path to portrait image
        portrait_url:
          type: string
          format: uri
          nullable: true
          description: URL to download portrait
        three_view:
          type: object
          properties:
            front:
              type: string
              nullable: true
            side:
              type: string
              nullable: true
            back:
              type: string
              nullable: true
    Character:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/CharacterRole'
        age:
          type: string
        description:
          $ref: '#/components/schemas/CharacterDescription'
        visual_tags:
          type: array
          items:
            type: string
        assets:
          $ref: '#/components/schemas/CharacterAssets'
...
    LocationAssets:
      type: object
      properties:
        reference:
          type: string
          nullable: true
          description: Path to reference image
        reference_url:
          type: string
          format: uri
          nullable: true
          description: URL to download reference
    Location:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/LocationType'
        description:
          type: string
        visual_tags:
          type: array
          items:
            type: string
        assets:
          $ref: '#/components/schemas/LocationAssets'
...
    ChapterSummary:
      type: object
      required: [id, number, title, status]
      properties:
        id:
          type: string
        number:
          type: integer
        title:
          type: string
        status:
          $ref: '#/components/schemas/ChapterStatus'
        scene_count:
          type: integer
        panel_count:
          type: integer
        generated_panel_count:
          type: integer
...
    Dialogue:
      type: object
      required: [text, type]
      properties:
        character_id:
          type: string
          nullable: true
        text:
          type: string
        type:
          $ref: '#/components/schemas/DialogueType'
    PanelCharacter:
      type: object
      required: [character_id]
      properties:
        character_id:
          type: string
        expression:
          type: string
          default: neutral
        pose:
          type: string
        position:
          type: string
          enum: [left, center, right, background]
          default: center
    PanelComposition:
      type: object
      properties:
        shot_type:
          $ref: '#/components/schemas/ShotType'
        angle:
          $ref: '#/components/schemas/CameraAngle'
        focus:
          type: string
    Panel:
      type: object
      required: [id, number]
      properties:
        id:
          type: string
        number:
          type: integer
        type:
          type: string
          enum: [panel, transition, splash]
          default: panel
        composition:
          $ref: '#/components/schemas/PanelComposition'
        characters:
          type: array
          items:
            $ref: '#/components/schemas/PanelCharacter'
        action:
          type: string
        dialogue:
          type: array
          items:
            $ref: '#/components/schemas/Dialogue'
        sfx:
          type: array
          items:
            type: string
        continues_from_previous:
          type: boolean
        continuity_note:
          type: string
        image_path:
          type: string
          nullable: true
        image_url:
          type: string
          format: uri
          nullable: true
    CreatePanelsRequest:
      type: object
      properties:
        style:
          type: string
          default: webtoon
          description: Art style for generation
...